Safari-611.1.10/LayoutTests/mathml/presentation/msub-subscriptshift.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
        <script>
            function runTest() {
                description('Tests that subscriptshift works on msub. After adding subscriptshift the height of msub element should increase');

                var rect = document.getElementById("msubEle").getBoundingClientRect();
                height = rect.bottom - rect.top;

                shouldBeTrue("height > 100");
                isSuccessfullyParsed();
            }
        </script>
    </head>
    <body onload="runTest()">        
        <math>
            <msub subscriptshift="60px" id="msubEle">
                <mspace width="50px" height="50px" mathbackground="green"/>
                <mspace width="50px" height="50px" mathbackground="blue"/>
            </msub>
        </math>
    </body>
</html>
