LayoutTests/mathml/presentation/msub-subscriptshift-expected.txt

Tests that subscriptshift works on msub. After adding subscriptshift the height of msub element should increase

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS height > 100 is true
PASS successfullyParsed is true

TEST COMPLETE
